requirements.txt
streamlit
opencv-python
pillow
numpy
dlib
moviepy
pyttsx3
